Ignore the Cordova plugin if it doesn't have platform tag
